+++ CHANGELOG.md
@@ -4,12 +4,16 @@ TODO
 ====
 * Add io values stack (when offline)
 * Add full MQTT OTA update (missing binary)
-* Add full HTTP PTA update (missing version)
+* Add full HTTP OTA update (missing version)
+
+OptaLinker 1.3.1 - Incoming
+===========================
+* Add button to web interface to reset ALL IO counters
 
 OptaLinker 1.3.0 - 2026.07.05
 =============================
 * Change MQTT library from ArduinoMQTTClient to MQTT (at https://github.com/256dpi/arduino-mqtt )
-* Update OT exemple
+* Update OTA exemples to 1.3.0
 
 OptaLinker 1.2.3 - 2026.06.11
 =============================